Migrar script de diagnóstico GPU a la carpeta src y eliminar archivo obsoleto.
