--- src/utils/gpu/gpu.py
@@ -1,21 +1,0 @@
-# This script checks if CUDA is available and selects the appropriate device (GPU or CPU) for performing tensor operations with PyTorch.
-# Additionally, it prints the selected device to the console.
-
-import torch
-
-def check_tensor_operations():
-    """
-    Checks the availability of CUDA and selects the appropriate device.
-
-    Returns:
-        torch.device: The selected device (GPU or CPU).
-    """
-    # Check if CUDA is available
-    if torch.cuda.is_available():
-        print("CUDA is available.\nTensor operations can be performed on GPU.")
-        device = torch.device("cuda")
-    else:
-        print("CUDA is not available.\nTensor operations will be performed on CPU.")
-        device = torch.device("cpu")
-    
-    return device
